% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brands.R
\docType{data}
\name{brands}
\alias{brands}
\title{Dummy data set for 1 SampleChi Square Proportion Test}
\format{A data frame with 43 rows and 2 variables:
\describe{
  \item{id}{index}
  \item{brand}{smartphone brand}
}}
\source{
\url{http://www.spss-tutorials.com/spss-one-sample-chi-square-test/}
}
\usage{
brands
}
\description{
A dataset containing information about smartphone brands.
}
\keyword{datasets}

