% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochran-q-test.R
\name{cochran_test}
\alias{cochran_test}
\title{Cochran Q Test}
\usage{
cochran_test(x, ...)
}
\arguments{
\item{x}{a data frame/vector}

\item{...}{numeric vectors}
}
\value{
\code{cochran_test} returns an object of class \code{"cochran_test"}.
An object of class \code{"cochran_test"} is a list containing the
following components:

\item{n}{number of observations}
\item{df}{degrees of freedom}
\item{q}{cochran's q statistic}
\item{pvalue}{p-value}
}
\description{
Test if the proportions of 3 or more dichotomous variables are
equal in the same population.
}
\examples{
cochran_test(exam)
}
\references{
Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric
Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
}

