% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-cochran-q-test.R
\name{infer_cochran_qtest}
\alias{infer_cochran_qtest}
\alias{cochran_test}
\title{Cochran Q Test}
\usage{
infer_cochran_qtest(x, ...)
}
\arguments{
\item{x}{a data frame/vector}

\item{...}{numeric vectors}
}
\value{
\code{infer_cochran_qtest} returns an object of class
\code{"infer_cochran_qtest"}. An object of class \code{"infer_cochran_qtest"}
is a list containing the following components:

\item{n}{number of observations}
\item{df}{degrees of freedom}
\item{q}{cochran's q statistic}
\item{pvalue}{p-value}
}
\description{
Test if the proportions of 3 or more dichotomous variables are
equal in the same population.
}
\section{Deprecated Function}{

\code{cochran_test()} has been deprecated. Instead use
\code{infer_cochran_qtest()}.
}

\examples{
infer_cochran_qtest(exam)
}
\references{
Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric
Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
}
