% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifr-launch-shiny-app.R
\name{infer_launch_shiny_app}
\alias{infer_launch_shiny_app}
\title{Launch Shiny App}
\usage{
infer_launch_shiny_app()
}
\description{
Launches shiny app
}
\examples{
\dontrun{
infer_launch_shiny_app()
}
}
