% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferr.R
\docType{package}
\name{inferr}
\alias{inferr}
\title{\code{inferr} package}
\description{
Parametric and non parametric statistical tests
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/inferr}{GitHub}
}
