% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene-test.R
\name{levene_test}
\alias{levene_test}
\alias{levene_test.formula}
\alias{levene_test.lm}
\title{Levene test for equality of variances}
\usage{
levene_test(variable, ..., group_var = NA, alternative = c("mean", "median",
  "trimmed-mean", "all"), trim.mean = 0.1)

\method{levene_test}{lm}(model)

\method{levene_test}{formula}(formula, data)
}
\arguments{
\item{variable}{a numeric vector}

\item{...}{numeric vectors}

\item{group_var}{a grouping variable}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "both" (default), "greater", "less" or "all". You can specify
just the initial letter.}
}
\value{
\code{levene_test} returns an object of class \code{"levene_test"}.
An object of class \code{"levene_test"} is a list containing the
following components:

\item{bf}{Brown and Forsythe f statistic}
\item{p_bf}{p-value for Brown and Forsythe f statistic}
\item{lev}{Levene's f statistic}
\item{p_lev}{p-value for Levene's f statistic}
\item{bft}{Brown and Forsythe f statistic using trimmed mean}
\item{p_bft}{p-value for Brown and Forsythe f statistic using trimmed mean}
\item{avgs}{mean for each level of the grouping variable}
\item{sds}{standard deviations for each level of the grouping variable}
\item{avg}{combined mean}
\item{sd}{combined standard deviation}
\item{n}{number of observations}
\item{n_df}{numerator degrees of freedom}
\item{d_df}{denominator degrees of freedom}
\item{levs}{levels of the grouping variable}
\item{lens}{number of observations for each level of the grouping variable}
\item{type}{alternative hypothesis}
}
\description{
\code{levene_test} reports Levene's robust test statistic (W_0)
for the equality of variances between the groups defined by groupvar and the
two statistics proposed by Brown and Forsythe that replace the mean in
Levene's formula with alternative location estimators. The first alternative
(W_50) replaces the mean with the median.  The second alternative replaces
the mean with the 10% trimmed mean (W_10).
}
\examples{
# using grouping variable
levene_test(mtcars$mpg, group_var = mtcars$cyl, alternative = 'mean')
levene_test(mtcars$mpg, group_var = mtcars$cyl, alternative = 'median')
levene_test(mtcars$mpg, group_var = mtcars$cyl, alternative = 'trimmed-mean')
levene_test(mtcars$mpg, group_var = mtcars$cyl, alternative = 'all')

# using two variables
levene_test(mtcars$mpg, mtcars$qsec, alternative = 'mean')
levene_test(mtcars$mpg, mtcars$qsec, alternative = 'median')
levene_test(mtcars$mpg, mtcars$qsec, alternative = 'trimmed-mean')
levene_test(mtcars$mpg, mtcars$qsec, alternative = 'all')

}

