% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-samp-chisq-prop-test.R
\name{os_chisqgof}
\alias{os_chisqgof}
\title{One Sample Chi Square Test}
\usage{
os_chisqgof(x, y, correct = FALSE, conf.int = 0.95)
}
\arguments{
\item{x}{categorical variable}

\item{y}{expected proportion/frequencies}

\item{correct}{logical; if TRUE continuity correction is applied}

\item{conf.int}{confidence level}
}
\value{
\code{os_chisqgof} returns an object of class \code{"os_chisqgof"}.
An object of class \code{"os_chisqgof"} is a list containing the
following components:

\item{chisquare}{chi square statistic}
\item{pvalue}{p-value}
\item{df}{chi square degrees of freedom}
\item{ssize}{number of observations}
\item{names}{levels of \code{x}}
\item{level}{number of levels of \code{x}}
\item{obs}{observed frequency/proportion}
\item{exp}{expected frequency/proportion}
\item{deviation}{deviation of observed from frequency}
\item{std}{standardized residuals}
\item{varname}{name of categorical variable}
\item{c_low}{lower confidence limit for mean}
\item{c_up}{upper confidence limit for mean}
}
\description{
Test whether a single categorical variable follows a
hypothesized population distribution.
}
\examples{
os_chisqgof(hsb$female, c(100, 100))
os_chisqgof(hsb$race, c(50, 50, 50, 50))

# continuity correction
os_chisqgof(hsb$female, c(100, 100), correct = TRUE)
}
\references{
Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric
Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
}
\seealso{
\code{\link[stats]{prop.test}} \code{\link[stats]{binom.test}}
}

