% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_test}
\alias{runs_test}
\title{Test for Random Order}
\usage{
runs_test(x, drop = FALSE, split = FALSE, mean = FALSE, threshold = NA)
}
\arguments{
\item{x}{numeric vector}

\item{drop}{logical; if TRUE, values equal to the threshold will be dropped
from \code{x}}

\item{split}{logical; if TRUE, data will be recoded in binary format}

\item{mean}{logical; if TRUE, mean will be used as threshold}

\item{threshold}{threshold to be used for counting runs, specify 0 if data
is coded as a binary.}
}
\value{
\code{runs_test} returns an object of class \code{"runs_test"}.
An object of class \code{"runs_test"} is a list containing the
following components:

\item{n}{number of observations}
\item{threshold}{within group sum of squares}
\item{n_below}{number below the threshold}
\item{n_above}{number above the threshold}
\item{mean}{expected number of runs}
\item{var}{variance of the number of runs}
\item{n_runs}{number of runs}
\item{z}{z statistic}
\item{p}{p-value of \code{z}}
}
\description{
runtest tests whether the observations of \code{x} are serially
independent -- that is, whether they occur in a random order -- by counting
how many runs there are above and below a threshold.  By default, the median
is used as the threshold.  A small number of runs indicates positive serial
correlation; a large number indicates negative serial correlation.
}
\examples{
reg <- lm(mpg ~ disp, data = mtcars)
runs_test(residuals(reg))
runs_test(residuals(reg), drop = TRUE)
runs_test(residuals(reg), split = TRUE)
runs_test(residuals(reg), mean = TRUE)
runs_test(residuals(reg), threshold = 0)
}

