% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment2.R
\docType{data}
\name{treatment2}
\alias{treatment2}
\title{Dummy data set for 2 Sample Proportion test}
\format{A data frame with 200 rows and 2 variables:
\describe{
  \item{outcome}{outcome of treatment}
  \item{female}{gender of patient, 0 for male and 1 for female}
}}
\usage{
treatment2
}
\description{
A dataset containing information about treatment outcomes
}
\keyword{datasets}
