% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifr-cochran-q-test.R
\name{infer_cochran_qtest}
\alias{infer_cochran_qtest}
\title{Cochran Q Test}
\usage{
infer_cochran_qtest(data, ...)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble}}

\item{...}{columns in \code{data}}
}
\value{
\code{infer_cochran_qtest} returns an object of class
\code{"infer_cochran_qtest"}. An object of class \code{"infer_cochran_qtest"}
is a list containing the following components:

\item{df}{degrees of freedom}
\item{n}{number of observations}
\item{pvalue}{p value}
\item{q}{cochran's q statistic}
}
\description{
Test if the proportions of 3 or more dichotomous variables are
equal in the same population.
}
\section{Deprecated Function}{

\code{cochran_test()} has been deprecated. Instead use
\code{infer_cochran_qtest()}.
}

\examples{
infer_cochran_qtest(exam, exam1, exam2, exam3)
}
\references{
Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric
Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
}
