% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferr.R
\docType{package}
\name{inferr}
\alias{inferr}
\alias{inferr-package}
\title{\code{inferr} package}
\description{
Parametric and non parametric statistical tests
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/inferr}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rsquaredacademy.github.io/inferr/}
  \item \url{https://github.com/rsquaredacademy/inferr}
  \item Report bugs at \url{https://github.com/rsquaredacademy/inferr/issues}
}

}
\author{
\strong{Maintainer}: Aravind Hebbali \email{hebbali.aravind@gmail.com} (\href{https://orcid.org/0000-0001-9220-9669}{ORCID})

}
\keyword{internal}
