% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar-test.R
\name{mcnemar_test}
\alias{mcnemar_test}
\title{McNemar Test}
\usage{
mcnemar_test(x, y = NULL)
}
\arguments{
\item{x}{2 x 2 matrix or 2 x 2 table or numeric variable or factor variable}

\item{y}{numeric or factor variable}
}
\value{
\code{mcnemar_test} returns an object of class \code{"mcnemar_test"}.
An object of class \code{"mcnemar_test"} is a list containing the
following components:

\item{statistic}{chi square statistic}
\item{df}{degrees of freedom}
\item{pvalue}{p-value}
\item{exactp}{exact p-value}
\item{cstat}{continuity correction chi square statistic}
\item{cpvalue}{continuity correction p-value}
\item{kappa}{kappa coefficient}
\item{std_err}{asymptotic standard error}
\item{kappa_cil}{95\% lower confidence limit}
\item{kappa_ciu}{95\%upper confidence limit}
\item{cases}{cases}
\item{controls}{controls}
\item{ratio}{ratio of proportion with factor}
\item{odratio}{odds ratio}
\item{tbl}{two way table}
}
\description{
Test if the proportions of two dichotomous variables are
equal in the same population.
}
\examples{
# test if the proportion of students in himath and hiread group is same
himath <- ifelse(hsb$math > 60, 1, 0)
hiread <- ifelse(hsb$read > 60, 1, 0)
mcnemar_test(table(himath, hiread))

# using matrix
mcnemar_test(matrix(c(135, 18, 21, 26), nrow = 2))
}
\references{
Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric
Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
}
\seealso{
\code{\link[stats]{mcnemar.test}}
}

