% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-samp-prop-test.R
\name{ts_prop_test}
\alias{ts_prop_calc}
\alias{ts_prop_grp}
\alias{ts_prop_test}
\title{Two sample test of proportion}
\usage{
ts_prop_test(var1, var2, alternative = c("both", "less", "greater", "all"))

ts_prop_grp(var, group, alternative = c("both", "less", "greater", "all"))

ts_prop_calc(n1, n2, p1, p2, alternative = c("both", "less", "greater",
  "all"))
}
\arguments{
\item{var1}{a categorical variable}

\item{var2}{a categorical variable}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "both" (default), "greater", "less" or "all". You can specify
just the initial letter.}

\item{var}{a categorical variable}

\item{group}{a categorical variable}

\item{n1}{sample 1 size}

\item{n2}{sample 2 size}

\item{p1}{sample 1 proportion}

\item{p2}{sample 2 proportion}
}
\value{
an object of class \code{"prop_test"}.
An object of class \code{"prop_test"} is a list containing the
following components:

\item{n1}{sample 1 size}
\item{n2}{sample 2 size}
\item{phat1}{sample 1 proportion}
\item{phat2}{sample 2 proportion}
\item{z}{z statistic}
\item{sig}{p-value for z statistic}
\item{alt}{alternative hypothesis}
}
\description{
tests on the equality of proportions using
large-sample statistics. It tests that \code{var} has the same proportion
within two groups defined by \code{group} or that two variables \code{var1} and
\code{var2} have the same proportion.
}
\examples{
# using variables
ts_prop_test(var1 = mtcars$am, var2 = mtcars$vs, alternative = 'less')
ts_prop_test(var1 = mtcars$am, var2 = mtcars$vs, alternative = 'greater')
ts_prop_test(var1 = mtcars$am, var2 = mtcars$vs, alternative = 'both')
ts_prop_test(var1 = mtcars$am, var2 = mtcars$vs, alternative = 'all')

# using groups
ts_prop_grp(var = mtcars$am, group = mtcars$vs, alternative = 'less')
ts_prop_grp(var = mtcars$am, group = mtcars$vs, alternative = 'greater')
ts_prop_grp(var = mtcars$am, group = mtcars$vs, alternative = 'both')
ts_prop_grp(var = mtcars$am, group = mtcars$vs, alternative = 'all')

# using sample size and proportions
ts_prop_calc(n1 = 30, n2 = 25, p1 = 0.3, p2 = 0.5, alternative = 'less')
ts_prop_calc(n1 = 30, n2 = 25, p1 = 0.3, p2 = 0.5, alternative = 'greater')
ts_prop_calc(n1 = 30, n2 = 25, p1 = 0.3, p2 = 0.5, alternative = 'both')
ts_prop_calc(n1 = 30, n2 = 25, p1 = 0.3, p2 = 0.5, alternative = 'all')
}
\references{
Sheskin, D. J. 2007. Handbook of Parametric and Nonparametric
Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
}
\seealso{
\code{\link[stats]{prop.test}}
}

